Refactor texture system
Add Renderbuffer.cs and Framebuffer.cs.

Add GLResource.cs and GLObject.cs for easy OpenGL state/memory management.

Add colored glass mod for testing.

Split chunk rendering shaders to opaque (no transparency), cutout (full transparency = alpha clip) and translucent (partial transparency).

+++ source/client/source/Configuration/Constants.cs
@@ -265,7 +265,7 @@ public static class Constants
     /// <summary>
     /// The level of anisotropic filtering to use.
     /// </summary>
-    public const float ANISOTROPIC_FILTERING_LEVEL = 16f;
+    public const int ANISOTROPIC_FILTERING_LEVEL = 16;
 
     #endregion
